requirements.txt
pandas
numpy
streamlit
matplotlib
openpyxl
seaborn
folium
streamlit-folium
plotly-express
geopandas
vega_datasets